+++ plugin/poppler/gvloadimage_poppler.c
@@ -122,7 +122,7 @@ static void gvloadimage_poppler_cairo(GVJ_t * job, usershape_t *us, boxf b, bool
 
         cairo_save(cr);
 
- 	surface = cairo_image_surface_create (CAIRO_FORMAT_ARGB32, us->w, us->h);
+ 	surface = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, (int)us->w, (int)us->h);
 	cairo_surface_reference(surface);
 
         cairo_set_source_surface(cr, surface, 0, 0);
